ast: move idl.Position to the ast package (#504)
